backend/requirements.txt
fastapi
uvicorn
python-dotenv
Pillow
numpy
cryptography
pydub
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[argon2]
argon2-cffi